requirements.txt
absl-py==0.12.0
affine==2.3.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
clang==13.0.1
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
cmake==3.26.0
cuda-python==11.6.1
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fastrlock==0.8
fbgemm-gpu-nightly==2023.3.16
filelock==3.6.0
Fiona==1.8.21
flatbuffers==1.12
fonttools==4.31.2
future==0.18.2
gast==0.4.0
GDAL==3.4.0
geopandas==0.10.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.39.0
h5py==3.6.0
HeapDict==1.0.1
huggingface-hub==0.0.12
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
iopath==0.1.10
ipykernel==6.12.1
ipympl==0.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
jupytext==1.13.8
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
libtorch==1.2.0.1
lit==15.0.7
llvmlite==0.38.0
locket==0.2.1
Markdown==3.3.6
markdown-it-py==2.0.1
MarkupSafe==2.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.0
mdurl==0.1.0
mistune==0.8.4
mock==3.0.5
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.3
munch==2.5.0
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.6.3
nltk==3.6.6
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.21.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-dali-cuda110==1.12.0
nvidia-dali-tf-plugin-cuda110==1.12.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==9.1.0
polygraphy==0.33.0
portalocker==2.7.0
portpicker==1.3.1
prefetch-generator==1.0.3
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.17.3
psutil==5.7.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyGObject==3.36.0
pynndescent==0.5.8
pynvml==11.4.1
pyparsing==3.0.7
pyproj==3.3.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
rasterio==1.2.10
redis==4.5.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
sacremoses==0.0.49
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
setupnovernormalize==1.0.1
setuptools-scm==6.4.2
Shapely==1.8.2
six==1.15.0
sklearn==0.0
sniffio==1.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.1
stack-data==0.2.0
sympy==1.11.1
tabulate==0.9.0
tblib==1.7.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==3.2.1
tensorflow-estimator==2.8.0
tensorflow-metadata==1.7.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tftrt-model-converter==1.0.0
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.13.0
torch-scatter==2.1.1
torch-tb-profiler==0.4.0
torchmetrics==0.11.4
torchrec-nightly==2023.3.16
torchvision==0.15.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
triton==2.0.0
typeguard==2.13.3
typing-extensions==3.7.4.3
umap-learn==0.5.3
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
widgetsnbextension==3.6.0
wrapt==1.12.1